requirements.txt
mock
pytest
flask
redis
mockredispy
